Function left shifts bits of a number by n times.

+++ C/shift/left_shift.c
@@ -8,7 +8,7 @@
  * Return: Number
 */
 
-int left_shift(int number, int times)
+int left_shift(long number, int times)
 {
 	number <<= times;
 	return number;
